relocate client
